Use the organisations API from whitehall instead of the need_api
The use of the need API was a bit of a hack, we should use the proper
endpoint.

This means whitehall will have to be running too.

Modified the `id` of the org to be the slug to maintain compatability
with the old need_api version of orgs

Also remove some code which was not used after the lruchache was implemented