Change the Dashboard to display content plans
The original idea was to use tags to drive the dashboard. After importing
the plans, it seems this is no longer desired user behavior (coupled with the deletion of tags on prod),
so we are changing it to display the status of each plan instead.

Content plans will not display if there is no content for the filter